pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "adr-agent"
version = "0.1.0"
description = "Agentic CLI tool for generating Architecture Decision Records"
requires-python = ">=3.11"
dependencies = [
    "python-dotenv>=1.0",
    "anthropic>=0.40",
    "openai>=1.30",
    "typer[all]>=0.12",
    "rich>=13",
    "pydantic>=2.6",
    "pyyaml>=6",
    "tavily-python>=0.3",
]

[project.scripts]
adr-agent = "src.cli:app"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]
